Added games & authors list in app file
