+++ game.rb
@@ -11,6 +11,11 @@ class Game < Item
   end
 
   def can_be_archived?()
-    super && ((Date.today.year - @last_played_at.year) > 2)
+    today = Time.now
+    total = today.year - date.year
+    total -= 1 if
+        date.month > today.month or
+        (date.month >= today.month and date.day > today.day)
+    super && (total > 2)
   end
 end
